+++ core/startup_stm32h750xx.c
@@ -1547,8 +1547,12 @@ void __attribute__((naked, noreturn)) Reset_Handler()
 	for (pDest = &_sbss; pDest != &_ebss; pDest++)
 		*pDest = 0;
 
+	#ifndef BOOT_APP
 	SystemInit();
+	#endif
 	__libc_init_array();
+
+
 	(void)main();
 	for (;;) ;
 }