+++ src/per/sai.h
@@ -112,6 +112,9 @@ class SaiHandle
     /** Initializes an SAI peripheral */
     Result Init(const Config& config);
 
+    /** Deinitializes an SAI peripheral */
+    Result DeInit();
+
     /** Returns the current configuration */
     const Config& GetConfig() const;